philippian_encoder/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==5.1.6
django-celery-beat==2.7.0
django-import-export==4.3.5
django-jazzmin==3.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
dnspython==2.7.0
eventlet==0.39.0
greenlet==3.1.1
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
psycopg==3.2.3
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
sentry-sdk==2.21.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
